feat(iota-indexer): use shared cluster approach and add feature gate
