+++ crates/iota-indexer/tests/rpc-tests/main.rs
@@ -1,9 +1,9 @@
 // Copyright (c) 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-#[cfg(feature = "pg_integration")]
+#[cfg(feature = "rpc_tests")]
 #[path = "../common/mod.rs"]
 mod common;
 
-#[cfg(feature = "pg_integration")]
+#[cfg(feature = "rpc_tests")]
 mod read_api;
